client/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-uniauth==1.0.4
docutils==0.14
idna==2.8
jmespath==0.9.4
lxml==4.3.2
PyJWT==1.7.1
python-cas==1.4.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
